requirements.txt
langchain-core
langchain-text-splitters
pymilvus
requests
pyyaml
